+++ src/main/java/nl/knaw/dans/dvcli/command/TruncateNotifications.java
@@ -22,7 +22,6 @@ import nl.knaw.dans.dvcli.action.ConsoleReport;
 import nl.knaw.dans.dvcli.action.Database;
 import nl.knaw.dans.dvcli.action.Pair;
 import nl.knaw.dans.dvcli.action.ThrowingFunction;
-
 import picocli.CommandLine;
 
 import java.io.IOException;
@@ -31,104 +30,101 @@ import java.util.ArrayList;
 import java.util.List;
 
 @CommandLine.Command(name = "truncate-notifications",
-        mixinStandardHelpOptions = true,
-        description = "Remove user notifications but keep up to a specified amount.",
-        sortOptions = false)
+                     mixinStandardHelpOptions = true,
+                     description = "Remove user notifications but keep up to a specified amount.",
+                     sortOptions = false)
 @Slf4j
-public class NotificationTruncate extends AbstractCmd {
+public class TruncateNotifications extends AbstractCmd {
     protected Database db;
-    public NotificationTruncate(@NonNull Database database) {
+
+    public TruncateNotifications(@NonNull Database database) {
         this.db = database;
     }
 
     private static final long DEFAULT_DELAY = 10L; // 10 ms, database should be able to handle this
-    
+
     @CommandLine.Option(names = { "-d", "--delay" }, description = "Delay in milliseconds between requests to the server (default: ${DEFAULT-VALUE}).", defaultValue = "" + DEFAULT_DELAY)
     protected long delay = DEFAULT_DELAY;
-    
+
     @CommandLine.ArgGroup(exclusive = true, multiplicity = "1")
     UserOptions users;
 
     static class UserOptions {
-        @CommandLine.Option(names = { "--user" }, required = true, 
-                description = "The user database id (a number) whose notifications to truncate.")
+        @CommandLine.Option(names = { "--user" }, required = true,
+                            description = "The user database id (a number) whose notifications to truncate.")
         int user; // a number, preventing accidental SQL injection
         // This id is visible for 'superusers' in the Dataverse Dashboard
-        
-        @CommandLine.Option(names = { "--all-users" }, required = true, 
-                description = "Truncate notifications for all users.")
+
+        @CommandLine.Option(names = { "--all-users" }, required = true,
+                            description = "Truncate notifications for all users.")
         boolean allUsers;
     }
-    
-    @CommandLine.Parameters(index = "0", paramLabel = "number-of-records-to-keep", 
-            description = "The number of notification records to keep.")
+
+    @CommandLine.Parameters(index = "0", paramLabel = "number-of-records-to-keep",
+                            description = "The number of notification records to keep.")
     private int numberOfRecordsToKeep;
-    
+
     private record NotificationTruncateParams(Database db, int userId, int numberOfRecordsToKeep) {
     }
 
-    private BatchProcessor.BatchProcessorBuilder<NotificationTruncate.NotificationTruncateParams, String> paramsBatchProcessorBuilder() throws IOException {
-        return BatchProcessor.<NotificationTruncate.NotificationTruncateParams, String> builder();
+    private BatchProcessor.BatchProcessorBuilder<TruncateNotifications.NotificationTruncateParams, String> paramsBatchProcessorBuilder() throws IOException {
+        return BatchProcessor.<TruncateNotifications.NotificationTruncateParams, String> builder();
     }
-    
-    private static class NotificationTruncateAction implements ThrowingFunction<NotificationTruncate.NotificationTruncateParams, String, Exception> {
+
+    private static class NotificationTruncateAction implements ThrowingFunction<TruncateNotifications.NotificationTruncateParams, String, Exception> {
 
         @Override
         public String apply(NotificationTruncateParams notificationTruncateParams) throws Exception {
-            // Dry-run; show what will be deleted
-            //List<List<String>> results = notificationTruncateParams.db.query(String.format("SELECT * FROM usernotification WHERE user_id = '%d' AND id NOT IN (SELECT id FROM usernotification WHERE user_id = '%d' ORDER BY senddate DESC LIMIT %d);", 
-            //        notificationTruncateParams.userId, notificationTruncateParams.userId, 
-            //        notificationTruncateParams.numberOfRecordsToKeep), true
-            //);
-            //return "Notifications for user " + notificationTruncateParams.userId + " that will be deleted: \n" 
-            //        + getResultsAsString(results);
-            
-            // Actually delete the notifications
             try {
                 log.info("Deleting notifications for user with id {}", notificationTruncateParams.userId);
-                int rowCount = notificationTruncateParams.db.update(String.format("DELETE FROM usernotification WHERE user_id = '%d' AND id NOT IN (SELECT id FROM usernotification WHERE user_id = '%d' ORDER BY senddate DESC LIMIT %d);",
+                int rowCount = notificationTruncateParams.db.update(
+                    String.format("DELETE FROM usernotification WHERE user_id = '%d' AND id NOT IN (SELECT id FROM usernotification WHERE user_id = '%d' ORDER BY senddate DESC LIMIT %d);",
                         notificationTruncateParams.userId, notificationTruncateParams.userId,
                         notificationTruncateParams.numberOfRecordsToKeep));
                 return "Deleted " + rowCount + " record(s) for user with id " + notificationTruncateParams.userId;
-            } catch (SQLException e) {
+            }
+            catch (SQLException e) {
                 throw new Exception("Error deleting notifications for user with id " + notificationTruncateParams.userId, e);
             }
         }
     }
-    
+
     @Override
     public void doCall() throws Exception {
         // validate input
         if (numberOfRecordsToKeep < 0) {
             throw new Exception("Number of records to keep must be a positive integer, now it was " + numberOfRecordsToKeep + ".");
         }
-        
+
         db.connect();
         try {
             paramsBatchProcessorBuilder()
-                    .labeledItems(getItems())
-                    .action(new NotificationTruncate.NotificationTruncateAction())
-                    .delay(delay)
-                    .report(new ConsoleReport<>())
-                    .build()
-                    .process();
-        } finally {
+                .labeledItems(getItems())
+                .action(new TruncateNotifications.NotificationTruncateAction())
+                .delay(delay)
+                .report(new ConsoleReport<>())
+                .build()
+                .process();
+        }
+        finally {
             db.close();
         }
     }
-    
+
     List<Pair<String, NotificationTruncateParams>> getItems() throws Exception {
         List<Pair<String, NotificationTruncateParams>> items = new ArrayList<>();
         try {
             if (users.allUsers) {
                 getUserIds(db).forEach(user_id -> items.add(new Pair<>(Integer.toString(user_id),
-                        new NotificationTruncateParams(db, user_id, numberOfRecordsToKeep))));
-            } else {
+                    new NotificationTruncateParams(db, user_id, numberOfRecordsToKeep))));
+            }
+            else {
                 // single user
                 items.add(new Pair<>(Integer.toString(users.user),
-                        new NotificationTruncateParams(db, users.user, numberOfRecordsToKeep)));
+                    new NotificationTruncateParams(db, users.user, numberOfRecordsToKeep)));
             }
-        } catch (SQLException e) {
+        }
+        catch (SQLException e) {
             throw new Exception("Error getting user ids: ", e);
         }
         return items;
@@ -139,7 +135,7 @@ public class NotificationTruncate extends AbstractCmd {
         List<Integer> users = new ArrayList<Integer>();
         // Could just get all users with notifications
         // String sql = "SELECT DISTINCT user_id FROM usernotification;";
-        // Instead we want only users with to many notifications
+        // Instead we want only users with too many notifications
         String sql = String.format("SELECT user_id FROM usernotification GROUP BY user_id HAVING COUNT(user_id) > %d;", numberOfRecordsToKeep);
         List<List<String>> results = db.query(sql);
         for (List<String> row : results) {
@@ -147,7 +143,7 @@ public class NotificationTruncate extends AbstractCmd {
         }
         return users;
     }
-    
+
     private static String getResultsAsString(List<List<String>> results) {
         // Note that the first row could be the column names
         StringBuilder sb = new StringBuilder();
