+++ src/main/java/nl/knaw/dans/dvcli/command/collection/CollectionGetContents.java
@@ -13,23 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.dvcli.command;
+package nl.knaw.dans.dvcli.command.collection;
 
+import nl.knaw.dans.dvcli.command.AbstractCmd;
 import nl.knaw.dans.lib.dataverse.DataverseException;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.ParentCommand;
 
 import java.io.IOException;
 
-@Command(name = "is-metadata-blocks-root",
+@Command(name = "get-contents",
          mixinStandardHelpOptions = true,
-         description = "Determine if a dataverse collection inherits its metadata blocks from its parent.")
-public class CollectionIsMetadataBlocksRoot extends AbstractCmd {
+         description = "Show contents of a dataverse collection.")
+public class CollectionGetContents extends AbstractCmd {
     @ParentCommand
     private CollectionCmd collectionCmd;
 
     @Override
     public void doCall() throws IOException, DataverseException {
-        collectionCmd.batchProcessor(c -> c.isMetadataBlocksRoot().getEnvelopeAsString()).process();
+        collectionCmd.batchProcessor(c -> c.getContents().getEnvelopeAsString()).process();
     }
 }
